Adds python migrator package.
This package is intended to be extended later for multiple providers. It
takes a source repo and a destination repo as arguments. Currently, it
lacks the ability to declare repo types so it's set up to do from stash
to bitbucket cloud.

Components:
1. command-line interface
2. stash api interface
3. bitbucket cloud api interface
4. stash to bitbucket migration controller
5. requirements.txt
6. setup.py

Adds .gitignore with exclusions for common python and the data dir.

Makefile for container testing.

Adds Dockerfile and docker-compose environment.

compose environment spins up the stash server and creates an interactive
container to exec into for testing the code.

Adds init.sh script for fresh repo pulls.

Cleans up unneeded comments and prints